Support single dset read/write with multi-dset api
HDF5 1.13.3 supports multi-dataset read write. Two main new api
interfaces are introduced, one for read and one for write.

This commit modifies the Log VOL implementation to work with new apis.
Note the following two:
1. Tested on test cases without calling H5Dread_multi or H5Dwrite_multi.
2. Testing H5Dread_multi and H5Dwrite_multi will be in future commits.
3. Need to further consider error handling.
4. Should optimize for the case when multilpe regular file's datasets
exists in H5Dread_multi/H5Dwrite_multi.

+++ src/H5VL_log_introspect.cpp
@@ -54,7 +54,7 @@ herr_t H5VL_log_introspect_get_conn_cls (void *obj,
  *
  *-------------------------------------------------------------------------
  */
-herr_t H5VL_log_introspect_get_cap_flags (const void *info, unsigned *cap_flags) {
+herr_t H5VL_log_introspect_get_cap_flags (const void *info, uint64_t *cap_flags) {
     herr_t err          = 0;
     H5VL_log_info_t *ip = (H5VL_log_info_t *)info;
